% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Generate basic psychometrics from data and key}
\usage{
analyze(test, key = NULL, create_key = FALSE, id = TRUE, d = 2,
  use = "pairwise.complete.obs", na_to_0 = TRUE)
}
\arguments{
\item{test}{A \code{data.frame} containing subjects as rows and items as 
columns.  If \code{id = TRUE}, the first column should be an id column.}

\item{key}{Provides information on how to score the test.  If \code{NULL}, 
the default, then test is assumed to be already scored.  Two formats are 
accepted for the key.  A \emph{simple key} can be provided as a named 
vector or single-row data frame, where the names are the names of the 
items and the values are the correct answer (scored 1, and all other 
responses scored 0).  If the scoring scheme is more complex, a \emph{full 
key} can be provided; this must be provided as a \code{data.frame} where 
the first column, \code{response}, contains all the possible responses to 
any item, and the subsequent columns correspond to each item with each 
value in that column representing that item's score for that row's 
response.  See \code{vignette("scoring")} for more help on setting up the 
key.}

\item{create_key}{For SCORED data only, creates a dummy key mapping input to
output values.  This allows the distractor analysis to be used for scored
data}

\item{id}{Is there an id column provided for the test? If \code{FALSE} an id
column is automatically created and added.}

\item{d}{Number of digits for summary outputs}

\item{use}{Which observations to use (e.g. with or without missing values)
for calculating difficulty and covariance-based measures?  See \code{use}
argument in \code{\link{cov}}}

\item{na_to_0}{When scoring, assume that NAs are scored as 0? Default is
TRUE}
}
\value{
Returns an \code{analyze} object with various psychometrics 
  calculated.  Primarily to be used with the \code{\link{report}} function 
  for viewing detailed output.
}
\description{
Returns object with basic psychometrics of a test from data and key
}
\details{
This takes a raw test and a key and creates an object that stores some basic
psychometrics for the test.
}

