% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dif}
\alias{dif}
\title{Perform DIF}
\usage{
dif(keyed_test, group, focal_name, dif_type = "both", ...)
}
\arguments{
\item{keyed_test}{A student (row) by item (columns) data frame of  numeric responses or of the raw
responses The first column in the data frame is an ID column.}

\item{group}{Vector of group membership, should be same length as \code{keyed_test}}

\item{focal_name}{Name of the focal group (string)}

\item{dif_type}{character, "mh", "logistic", or "both". Default is "both"}

\item{\dots}{Passed onto difMH and/or difLogistic}
}
\description{
Wrapper function to perform Mantel-Hanzel or Logistic DIF
}
\details{
This function wraps the \code{difMH} and \code{difLogistic} from the \code{difR} package
}

