% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_analysis.R
\name{distractor_analysis}
\alias{distractor_analysis}
\title{Perform distractor analysis for QMEtest object}
\usage{
distractor_analysis(testQME, ...)
}
\arguments{
\item{x}{a \code{QMEtest} object}
}
\value{
A list with various aspects of the distractor analysis
}
\description{
Calculates distractor analysis for QMEtest objects
}

