% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distractor_report.R
\name{distractor_report}
\alias{distractor_report}
\title{Distractor report}
\usage{
distractor_report(x)
}
\arguments{
\item{x}{An \code{\link{analyze}} object}
}
\value{
As a side effect, prints distractor plots and distractor tables
}
\description{
Outputs distractor analysis and plots for \code{\link{analyze}} objects, intended to be used within \code{rmarkdown} documents
}
\details{
For each item, produces distractor plots and a table showing distractor performance.
}

