\docType{data}
\name{Graham}
\alias{Graham}
\title{Graham's (2006) data set number 1}
\arguments{
  \item{Graham}{variance-covariance matrix from item responses}
}
\description{
  Graham's (2006) data set (table) number 1 contains variances and covariances based on responses by N = 60 persons to a mix of tau-equivalent and tau-inequivalent items.
}
\examples{
data(Graham)
Graham
}
\author{
  Christopher T. Moore \email{moor0554@umn.edu}
}
\references{
  Graham, J. M. (2006). Congeneric and (essentially) tau-equivalent estimates of score reliability: What they are and how to use them. \emph{Educational and Psychological Measurement, 66}, 6, 930-944.
}
\keyword{data}
\keyword{datasets}
