% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_key.R
\name{is_valid_simple_key}
\alias{is_valid_full_key}
\alias{is_valid_simple_key}
\alias{valid_key}
\title{Utility functions to test validity of key.}
\usage{
is_valid_simple_key(key)
}
\arguments{
\item{key}{The candidate key \code{data.frame} or vector. See \code{vignette("scoring")}.}
}
\value{
\code{TRUE} or \code{false} depending on whether key is of relevant valid form
}
\description{
Test whether key is in valid format for scoring
}

