% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_full_key}
\alias{math_full_key}
\title{Full key for \code{math} dataset}
\format{A data frame with the first column, \code{response}, containing every
  possible response, and 10 additional columns specifying how each response
  will be scored for each item.}
\usage{
math_full_key
}
\description{
Contains correct answers for \code{math} dataset, in the \emph{full key} 
format required by \code{analyze}. See \code{vignette("scoring")}.
}
\seealso{
\code{\link{math}}, \code{\link{math_key}}
}
\keyword{datasets}

