% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_key}
\alias{math_key}
\title{Simple key for \code{math} dataset}
\format{A length 10 character vector containing the correct response for each of the 10
  items in \code{math}.}
\usage{
math_key
}
\description{
Contains correct answers for \code{math} dataset, in the \emph{simple key} 
format required by \code{analyze}. See \code{vignette("scoring")}.
}
\seealso{
\code{\link{math}}, \code{\link{math_full_key}}
}
\keyword{datasets}

