% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psycho_report.R
\name{psycho_report}
\alias{psycho_report}
\title{Create a test report}
\usage{
psycho_report(x, output = "output.html", open = TRUE)
}
\arguments{
\item{x}{An \code{odin_zeus} object}

\item{output}{The file name of the html report}

\item{open}{Whether to open the html report}
}
\value{
As a side effect, creates the html report
}
\description{
This function creates a report with the basic psychometrics
}
\details{
This reports contains number of items, number of examinees, a distribution of
the total scores, reliability estimates (Coefficient Alpha, Guttman's L2,
Guttman's L4, Feldt-Gilmer, and Feldt-Brennan), Item overview (difficulty,
point biserial correlation, corrected point biserial correlation, number of
missing responses, and probability of missing responses), and item details
}

