\docType{package}
\name{QME-package}
\alias{QME}
\alias{QME-package}
\title{Quantitative Methods in Education R package}
\description{
  This package contains functions written by students, faculty, and alumni of the Quantitative Methods in Education program at the University of Minnesota.
}
\details{
  This package contains functions for computing many different reliabilities, such as split-half.  
  It also contains a set of functions for estimating variance
  components, generalizability (G) coefficients, and
  universe scores when observed scores contain variation
  from one or more admissible measurement facets (e.g.,
  items and raters).  The set only applies univariate G theory
  at this time.
}
\seealso{
  \url{http://www.cehd.umn.edu/EdPsych/programs/QME/}
}
\keyword{package}

