% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QMEtest.R
\name{QMEtest}
\alias{QMEtest}
\title{Create QMEtest objects}
\usage{
QMEtest(test, key, id = TRUE)
}
\arguments{
\item{test}{A \code{data.frame} containing the test responses.  If \code{key}
is not provided, the test is assumed to be already scored. The data frame
should be formatted in the wide format, so that each row represents a
student, and each column represents an item. If \code{id = TRUE}, the
default, the first column is assumed to be the id.}

\item{key}{A vector or single-row \code{data.frame} indicating the keyed
responses (i.e. the correct answers for cognitive tests).}

\item{id}{Whether an ID column has been provided.  If \code{id = FALSE}, an id
 will be created in the output object.}
}
\value{
A \code{QMEtest} object. This is a list with the following elements:
 \describe{ \item{\code{raw_test}}{The raw test responses (\code{NULL} if the test is already scored).}
 \item{\code{key}}{The key}
 \item{\code{keyed_test}}{The keyed/scored test, using the \code{right_wrong} function}
 }
}
\description{
Creates a \code{QMEtest} object from given test responses and (optional) key
}

