% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Calculate reliability for QMEtest object}
\usage{
reliability(x, ...)
}
\arguments{
\item{x}{a \code{QMEtest} object}
}
\value{
A data frame with coefficent alpha, Guttman's L2 and L4,
  Feldt-Gilmer, and Feldt-Brennan reliabilities, CI limits and SEM
}
\description{
Calculates reliability functions for QMEtest objects
}

