% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{psycho_report}
\alias{report}
\title{Create a test report}
\usage{
report(x, report_filename = "QME_report", output_format = "html_document",
  output_dir = NULL, simple_html = FALSE, ...)
}
\arguments{
\item{x}{An \code{analyze} object}

\item{report_filename}{The file name of the report (not including file extension, which is determined automatically)}

\item{output_format}{One of \code{'html_document'}, \code{'word_document'},
or \code{'pdf_document'}.  Note that \code{'pdf_document'} requires a
working LaTeX implementation.}

\item{output_dir}{The path to the directory that the report will be written to.  If NULL, written to current working directory.}

\item{simple_html}{If \code{TRUE}, overrides \code{output_format} and outputs a character object containing a simple (R Markdown v1) html report, useful for embedding in Shiny apps.}

\item{...}{passed to \code{\link{render}}}
}
\value{
If \code{simple_html = FALSE}, the default, generates report file as a side effect.  If \code{simple_html = TRUE}, returns a string with the html of the report.
}
\description{
This function creates a report with the basic psychometrics
}
\details{
This reports contains number of items, number of examinees, a distribution of
the total scores, reliability estimates (Coefficient Alpha, Guttman's L2, 
Guttman's L4, Feldt-Gilmer, and Feldt-Brennan), item overview (difficulty, 
point-biserial correlation, corrected point-biserial correlation, number of 
missing responses, and probability of missing responses), and distractor analysis.

Rendering is done by means of \code{\link{render}} from the \code{rmarkdown}
package.  See that help page for more details on the rendering process.
}

