% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef_alpha.R
\name{coef_alpha}
\alias{coef_alpha}
\title{Coefficient Alpha}
\usage{
coef_alpha(x, ...)
}
\arguments{
\item{x}{is a data frame or matrix consisting of keyed item responses (0s and 1s).}
}
\value{
This function returns the value of coefficient alpha, the
lower and upper limits of Feldt's (1965) method, and standard error of
measurement for coefficient alpha.
}
\description{
This function computes coefficient alpha using the covariance matrix.
}
\examples{
library(QME)
data(math,math_key)
out = QMEtest(math, math_key)
x = getKeyedTestNoID(out)
coef_alpha(x)
}

