% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item_level.R
\name{item_level}
\alias{item_level}
\title{Perform item level analysis for QMEtest object}
\usage{
item_level(testQME, ...)
}
\arguments{
\item{x}{a \code{QMEtest} object}
}
\value{
A list with various aspects of the item level analysis
}
\description{
Calculates item level analysis for QMEtest objects
}

