% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Calculate reliability for QMEtest object}
\usage{
reliability(x, ...)
}
\arguments{
\item{x}{a \code{QMEtest} object}
}
\value{
A list with a bunch of tasty reliabilities
}
\description{
Calculates reliability functions for QMEtest objects
}

