\docType{data}
\name{Shavelson.Webb.5.1}
\alias{Shavelson.Webb.5.1}
\title{Shavelson and Webb's (1991) data set number 5.1}
\arguments{
  \item{Shavelson.Webb.5.1}{data frame in long format with
  a column for item scores and columns for sources of
  variance}
}
\description{
  Shavelson and Webb's (1991) data set (table) number 5.1
  contains item scores from a two-facet person * rater *
  subject generalizability (G) study with subject
  considered fixed.  Note that the data are in long format
  (one item score per row) and that the factors are
  classified as such.
}
\examples{
data(Shavelson.Webb.5.1)
head(Shavelson.Webb.5.1)
sapply(Shavelson.Webb.5.1, class)
}
\author{
  Christopher T. Moore \email{moor0554@umn.edu}
}
\references{
  Shavelson, R. J., & Webb, N. M. (1991).
  \emph{Generalizability theory: A primer}. New York: Sage
  Publications.
}
\keyword{data}
\keyword{datasets}

